Update for 2023 (#48)
Lots of changes for 2023, mostly to make things work with a new version of Spack, but also for now to detach from using tcl environment-modules, to just using spack load.

* Update to match current versions
* pre-commit fix
* Fix container section to actually work

Changes to the spack.yaml meant it didn't build right.  Also, the inner MPI
doesn't match the outer MPI, so running across multiple nodes won't work with
that example.  I've tidied it so it at least works on a single node.